created app entities and repository

+++ src/main/resources/db/migration/V1_0__create_tables.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS `revenue` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `location` int,
     `date` date,
     `sum` bigint
 );
 
 CREATE TABLE IF NOT EXISTS `customer` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `first_name` varchar(128),
     `last_name` varchar(128),
     `username` varchar(128),
@@ -15,18 +15,18 @@ CREATE TABLE IF NOT EXISTS `customer` (
 );
 
 CREATE TABLE IF NOT EXISTS `supplier` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(128)
 );
 
 CREATE TABLE IF NOT EXISTS `product_category` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(128),
     `description` varchar(128)
 );
 
 CREATE TABLE IF NOT EXISTS `order` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `shipped_from` int,
     `customer` int,
     `created_at` datetime,
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS `order` (
 );
 
 CREATE TABLE IF NOT EXISTS `location` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(128),
     `address_country` varchar(128),
     `address_city` varchar(128),
@@ -46,21 +46,21 @@ CREATE TABLE IF NOT EXISTS `location` (
 );
 
 CREATE TABLE IF NOT EXISTS `stock` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `product` int,
     `location` int,
     `quantity` int
 );
 
 CREATE TABLE IF NOT EXISTS `order_detail` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `order` int,
     `product` int,
     `quantity` int
 );
 
 CREATE TABLE IF NOT EXISTS `product` (
-    `id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(128),
     `description` varchar(128),
     `price` bigint,
